requirements.txt
realtimestt==0.3.104
fastapi
uvicorn[standard]
numpy
scipy
python-dotenv
